chore: update CI node version

+++ scripts/deploy.sh
@@ -10,23 +10,25 @@ REPO_URL="gitlab.com:dna-components/dna-library.git"
 
 # build distribution scripts
 npm run build
-# clone distribution repo
-rm -rf $REPO_DIR
-git clone $REPO_URL $REPO_DIR
-# empty lib dir
-rm -rf $DIST_DIR
-# copy distribution files
-cp -R $BUILD_DIR $DIST_DIR
-# update distribution project
-if [ -d "$REPO_DIR" ]; then
-    cd $REPO_DIR
-    sed -i "" "s/\"version\": \".*\"/\"version\": \"${TAG:1}\"/" package.json
-    sed -i "" "s/\"version\": \".*\"/\"version\": \"${TAG:1}\"/" bower.json
-    git add .
-    git commit -m "release: ${TAG}"
-    git tag -a ${TAG} -m "release: ${TAG}"
-    git push
-    npm publish
-    cd ..
+if [ -d "$BUILD_DIR" ]; then
+    # clone distribution repo
     rm -rf $REPO_DIR
+    git clone $REPO_URL $REPO_DIR
+    # empty lib dir
+    rm -rf $DIST_DIR
+    # copy distribution files
+    cp -R $BUILD_DIR $DIST_DIR
+    # update distribution project
+    if [ -d "$REPO_DIR" ]; then
+        cd $REPO_DIR
+        sed -i "" "s/\"version\": \".*\"/\"version\": \"${TAG:1}\"/" package.json
+        sed -i "" "s/\"version\": \".*\"/\"version\": \"${TAG:1}\"/" bower.json
+        git add .
+        git commit -m "release: ${TAG}"
+        git tag -a ${TAG} -m "release: ${TAG}"
+        git push
+        npm publish
+        cd ..
+        rm -rf $REPO_DIR
+    fi
 fi
